Move default screens from seeder to upgrade script
