+++ upgrades/2023_07_18_168972_add_default_screens_for_tasks.php
@@ -1,16 +1,19 @@
 <?php
 
-use Illuminate\Support\Facades\Artisan;
+use ProcessMaker\Models\Screen;
 use ProcessMaker\Upgrades\UpgradeMigration as Upgrade;
 
 class AddDefaultScreensForTasks extends Upgrade
 {
+    const SCREENS_PATH = 'processes/screens/';
+
     /**
      * Run the migrations.
      */
     public function up(): void
     {
-        Artisan::call('db:seed', ['--class' => 'ScreenSystemSeeder']);
+        $this->installScreen('default-display-screen');
+        $this->installScreen('default-form-screen');
     }
 
     /**
@@ -18,6 +21,23 @@ class AddDefaultScreensForTasks extends Upgrade
      */
     public function down(): void
     {
-        //
+    }
+
+    /**
+     * Install a screen.
+     */
+    private function installScreen(string $key): Screen
+    {
+        $path = database_path(self::SCREENS_PATH . "{$key}.json");
+        $json = json_decode(file_get_contents($path), true);
+        $screen = collect($json['screens'])->first();
+
+        // By default, the screen is initially installed as a system screen.
+        unset($screen['categories']);
+        $screen['screen_category_id'] = null;
+
+        return Screen::updateOrCreate([
+            'key' => $screen['key'],
+        ], $screen);
     }
 }
